Bugfix: thumbbit handling for PandaTarget
With the pretender additions, Qemu emulates the armv7m-isa now
correctly.
As side effect, the thumbbit in the cpsr is at a different location.
However, the PandaTarget is not in sync with these changes, hence we
handle the setting of thumbbit differently in the according callback in
archs/arm.py

This commit also includes a test for the CI to check that this change
works as intended.