Dev/v1.4.2 (#77)
* allow inlining of pyperipherals into qemu via API

* unicorn_caller plugin

* wrapper for add_hook

* adjust unicorn_caller import path

* Flags for transfer state rather than absolute bools

* allow avatar-reference to be dictifyable

* added overwrite option to memory ranges

* working version of mr-overwriting

* bugfix for overwrite option

* Rename memory region on overwrite to shorter name

QEMU has a memory name limit of 255. Memory regions with many splits can
create names that exceed this, leading to conflicts.

* additional debugging output

* Add pypanda add_hook function and testcase

The CI test is currently not working due to a bug on PANDAs hooking
infrastructure, which is investigated in parallel. Once fixed, the CI
test should work as expected.

* Bump version and updated changelog

Co-authored-by: Grant Hernandez <grant.h.hernandez@gmail.com>