targets/openocd: pass along multiple scripts
The OpenOCD protocol supports passing multiple scripts if a list is
passed. However, the OpenOCD target checked the path to the script using
os.path.exists(), which does not handle lists of paths.

We now check whether the path is a string or an iterable, and pass that
along. While we're there, we also upgrade the in protocols/openocd.py
to handle more generic iterables instead of `list`.