OpenOCDTarget: don't check for script existence
The OpenOCD protocol supports passing multiple scripts if a list is
passed. However, the OpenOCD target checked the path to the script using
`os.path.exists`, which does not handle lists of paths.

Furthermore, OpenOCD can search for scripts in other locations that
avatar2 does not have knowledge about, for example if a search path is
passed using `-s` in `additional_args`, or the global scripts path.