fixing a parsing error
The parsing assumes a 4 line header, which is removed and assumes to blank line.
However, it appears to have blank lines in the output, which causes index out of range errors. 
Solution: 1) splitting on the last part of the header "objfile" to remove unwanted header lines, and then ignoring the blank lines while parsing.